requirements.txt
Django==3.1.5
djangorestframework==3.12.2
django-environ==0.4.5
psycopg2==2.8.6
requests==2.25.1
django-cors-headers==3.7.0